requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.6.1
python-multipart==0.0.6
pandas==2.2.0
openpyxl==3.1.2
gspread==5.12.4
oauth2client==4.1.3 